refac(91026): Extrai task gerar_previa_relatorio_acertos_async
